cleaning up READMEs, renaming a bit
