+++ angular-client/src/environment/environment.ts
@@ -1,6 +1,5 @@
 export const environment = {
-  production: true,
-  url: 'http://192.168.100.1:8000',
+  production: false,
   mapbox: {
     accessToken: 'pk.eyJ1IjoibWNrZWVwIiwiYSI6ImNscXBrcmU1ZTBscWIya284cDFyYjR3Nm8ifQ.6TQHlxhAJzptZyV-W28dnw'
   }
