+++ package.json
@@ -11,12 +11,14 @@
     "prisma:generate": "cd scylla-server && npm run prisma:generate",
     "prisma:migrate": "cd scylla-server && npm run prisma:migrate",
     "prisma:migrate:prod": "cd scylla-server && npm run prisma:migrate:prod",
-    "build:proto": "cd scylla-server && npm run build:proto"
+    "build:proto": "cd scylla-server && npm run build:proto",
+    "start": "concurrently \"npm run start --prefix angular-client\" \"npm run start --prefix scylla-server\""
   },
   "author": "Northeastern Electric Racing",
   "license": "ISC",
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^6.7.0",
+    "concurrently": "^8.2.2",
     "eslint": "^8.49.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-config-react-app": "^7.0.1",
